ecommercev1/requirements.txt
djangorestframework==3.11.2
djangorestframework-simplejwt==4.7.1
django-extensions==3.1.3

# DRF and JWT specific packages
Django==3.1.13
djangorestframework==3.12.4
PyJWT==2.2.0
